requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
top_100==0.1.4